refactor: unify SDR freq/OFDM cfg numeric types
- Change `tx_freq`/`rx_freq`/`sample_rate` from `double` to `float` in `SDRConfig`
- Change `n_subcarriers`/`pilot_spacing`/`n_cp` from int to `size_t` in `OFDMConfig`
- Update `set_rx_freq`/`set_tx_freq`/`get_sample_rate` signatures accordingly
- Fix division literals to use `1e9f` instead of `1e9`

+++ include/common.hpp
@@ -47,9 +47,9 @@ struct DSP {
     float sample_rate = 1.92e6;
     struct OFDMConfig {
         Modulation mod = Modulation::QAM16;
-        int n_subcarriers = 128;
-        int pilot_spacing = 19;
-        int n_cp = 32;
+        size_t n_subcarriers = 128;
+        size_t pilot_spacing = 19;
+        size_t n_cp = 32;
     } ofdm_cfg;
 };
